requirements.txt
Flask==2.3.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gunicorn==21.2.0; sys_platform != "win32"
waitress==2.1.2; sys_platform == "win32"
python-dotenv==1.0.0
